Separate FP8 FusedMoE into separate module
